requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
pandas>=1.5.0
Pillow>=9.0.0
tqdm>=4.64.0
numpy>=1.21.0
scikit-learn>=1.1.0